refactor(stella-tools,stella-store): split the latency budget into its own module and move the repair-fold test beside the projection
Both file-size failures this change caused are fixed by splitting rather
than by touching the baseline.

Closes #4033
Closes #4034
Closes #4035
