+++ crates/stella-store/src/tests.rs
@@ -508,90 +508,6 @@ fn producer_materializes_tool_calls_reflection_and_rolls_up_to_usage() {
 }
 
 #[test]
-fn materialize_keeps_two_announcements_sharing_a_call_id_apart() {
-    use stella_protocol::{ToolCall, ToolOutput};
-    let store = Store::in_memory().unwrap();
-    let id = store
-        .begin_execution("deck", "add a feature", "zai", "glm-5.2")
-        .unwrap();
-
-    // The same call_id announced twice by two different events: two calls.
-    // `call_id` is unique only within one model response, so the repair fold
-    // must key on the announcing event and keep them apart — folding them
-    // erased 12.2% of one workspace's calls (#4033).
-    store
-        .record_event(
-            id,
-            0,
-            &AgentEvent::ToolStart {
-                call: ToolCall {
-                    call_id: "c1".into(),
-                    name: "grep".into(),
-                    input: serde_json::json!({"pattern": "first"}),
-                },
-            },
-        )
-        .unwrap();
-    store
-        .record_event(
-            id,
-            1,
-            &AgentEvent::ToolStart {
-                call: ToolCall {
-                    call_id: "c1".into(),
-                    name: "grep".into(),
-                    input: serde_json::json!({"pattern": "final"}),
-                },
-            },
-        )
-        .unwrap();
-    store
-        .record_event(
-            id,
-            2,
-            &AgentEvent::ToolResult {
-                call_id: "c1".into(),
-                output: ToolOutput::Ok {
-                    content: "hit\n".into(),
-                    data: None,
-                },
-                duration_ms: 12,
-                speculated: false,
-            },
-        )
-        .unwrap();
-
-    let n = store.materialize_tool_calls(id).unwrap();
-    assert_eq!(n, 2, "two announcements are two calls (#4033)");
-    assert_eq!(store.count("tool_calls").unwrap(), 2);
-    let calls: Vec<(String, i64, String)> = {
-        let conn = store.lock();
-        let mut stmt = conn
-            .prepare(
-                "SELECT args_json, ok, state FROM tool_calls \
-                 WHERE execution_id = ?1 ORDER BY seq ASC",
-            )
-            .unwrap();
-        let mapped = stmt
-            .query_map(params![id], |r| Ok((r.get(0)?, r.get(1)?, r.get(2)?)))
-            .unwrap();
-        mapped.map(|r| r.unwrap()).collect()
-    };
-    assert_eq!(
-        calls[0].0, r#"{"pattern":"first"}"#,
-        "each announcement keeps its own arguments"
-    );
-    assert_eq!(calls[1].0, r#"{"pattern":"final"}"#);
-    // Only one result was delivered, and it settles the older open call; the
-    // younger one is still outstanding, which is what abandonment means.
-    assert_eq!(calls[0].1, 1, "the result attached to the call it answers");
-    assert_eq!(
-        calls[1].2, "abandoned",
-        "an unanswered announcement is not silently merged away"
-    );
-}
-
-#[test]
 fn reflection_truncation_flag_requires_the_output_token_limit_marker() {
     let store = Store::in_memory().unwrap();
