+++ crates/stella-store/src/tool_calls/tests.rs
@@ -266,9 +266,7 @@ fn two_steps_announcing_one_call_id_project_two_rows() {
     let args: Vec<String> = {
         let conn = store.lock();
         let mut stmt = conn
-            .prepare(
-                "SELECT args_json FROM tool_calls WHERE execution_id = ?1 ORDER BY seq ASC",
-            )
+            .prepare("SELECT args_json FROM tool_calls WHERE execution_id = ?1 ORDER BY seq ASC")
             .expect("prepare");
         let mapped = stmt
             .query_map(params![id], |r| r.get::<_, String>(0))
@@ -281,9 +279,7 @@ fn two_steps_announcing_one_call_id_project_two_rows() {
     let sizes: Vec<i64> = {
         let conn = store.lock();
         let mut stmt = conn
-            .prepare(
-                "SELECT bytes_out FROM tool_calls WHERE execution_id = ?1 ORDER BY seq ASC",
-            )
+            .prepare("SELECT bytes_out FROM tool_calls WHERE execution_id = ?1 ORDER BY seq ASC")
             .expect("prepare");
         let mapped = stmt
             .query_map(params![id], |r| r.get::<_, i64>(0))
@@ -303,13 +299,25 @@ fn two_steps_announcing_one_call_id_project_two_rows() {
 #[test]
 fn duplicate_ids_within_one_response_stay_two_rows() {
     let (store, id) = fixture();
-    store.record_event(id, 0, &start("dup", "read_file")).unwrap();
-    store.record_event(id, 1, &start("dup", "read_file")).unwrap();
-    store.record_event(id, 2, &ok_result("dup", "first", 3)).unwrap();
-    store.record_event(id, 3, &err_result("dup", "second")).unwrap();
+    store
+        .record_event(id, 0, &start("dup", "read_file"))
+        .unwrap();
+    store
+        .record_event(id, 1, &start("dup", "read_file"))
+        .unwrap();
+    store
+        .record_event(id, 2, &ok_result("dup", "first", 3))
+        .unwrap();
+    store
+        .record_event(id, 3, &err_result("dup", "second"))
+        .unwrap();
 
     let folded = rows(&store, id);
-    assert_eq!(folded.len(), 2, "neither duplicate absorbed the other: {folded:?}");
+    assert_eq!(
+        folded.len(),
+        2,
+        "neither duplicate absorbed the other: {folded:?}"
+    );
     assert_eq!(folded[0].2, "ok", "the first result settled the older call");
     assert_eq!(folded[1].2, "error", "the second settled the younger");
 }
@@ -555,5 +563,92 @@ fn rollup_bucket_counts_errors_but_not_abandonment() {
         (bucket.calls, bucket.errors),
         (2, 1),
         "abandonment is a fact about the turn, not the tool (#3146)"
+    );
+}
+
+/// The repair fold keeps them apart too, or a turn-end re-materialization
+/// would silently re-collapse what the live fold recorded correctly. Moved
+/// here from `store::tests` with #4033: it is a projection test, and it
+/// belongs beside the projection.
+#[test]
+fn materialize_keeps_two_announcements_sharing_a_call_id_apart() {
+    let store = Store::in_memory().unwrap();
+    let id = store
+        .begin_execution("deck", "add a feature", "zai", "glm-5.2")
+        .unwrap();
+
+    // The same call_id announced twice by two different events: two calls.
+    // `call_id` is unique only within one model response, so the repair fold
+    // must key on the announcing event and keep them apart — folding them
+    // erased 12.2% of one workspace's calls (#4033).
+    store
+        .record_event(
+            id,
+            0,
+            &AgentEvent::ToolStart {
+                call: ToolCall {
+                    call_id: "c1".into(),
+                    name: "grep".into(),
+                    input: serde_json::json!({"pattern": "first"}),
+                },
+            },
+        )
+        .unwrap();
+    store
+        .record_event(
+            id,
+            1,
+            &AgentEvent::ToolStart {
+                call: ToolCall {
+                    call_id: "c1".into(),
+                    name: "grep".into(),
+                    input: serde_json::json!({"pattern": "final"}),
+                },
+            },
+        )
+        .unwrap();
+    store
+        .record_event(
+            id,
+            2,
+            &AgentEvent::ToolResult {
+                call_id: "c1".into(),
+                output: ToolOutput::Ok {
+                    content: "hit\n".into(),
+                    data: None,
+                },
+                duration_ms: 12,
+                speculated: false,
+            },
+        )
+        .unwrap();
+
+    let n = store.materialize_tool_calls(id).unwrap();
+    assert_eq!(n, 2, "two announcements are two calls (#4033)");
+    assert_eq!(store.count("tool_calls").unwrap(), 2);
+    let calls: Vec<(String, i64, String)> = {
+        let conn = store.lock();
+        let mut stmt = conn
+            .prepare(
+                "SELECT args_json, ok, state FROM tool_calls \
+                 WHERE execution_id = ?1 ORDER BY seq ASC",
+            )
+            .unwrap();
+        let mapped = stmt
+            .query_map(params![id], |r| Ok((r.get(0)?, r.get(1)?, r.get(2)?)))
+            .unwrap();
+        mapped.map(|r| r.unwrap()).collect()
+    };
+    assert_eq!(
+        calls[0].0, r#"{"pattern":"first"}"#,
+        "each announcement keeps its own arguments"
+    );
+    assert_eq!(calls[1].0, r#"{"pattern":"final"}"#);
+    // Only one result was delivered, and it settles the older open call; the
+    // younger one is still outstanding, which is what abandonment means.
+    assert_eq!(calls[0].1, 1, "the result attached to the call it answers");
+    assert_eq!(
+        calls[1].2, "abandoned",
+        "an unanswered announcement is not silently merged away"
     );
 }
